Refactor panels and global stores
This refactors the panels a bit and makes them look a bit prettier.

It also refactors the global stores (stores.ts) into a single global object (now in $lib/state.svelte.ts).

Finally, this reorganizes the CSS a bit, so the global CSS file (now combined with app.css) is actually global.